% tinyrox says don't edit this manually, but it can't stop you!
\name{apply_interleaved_rotary_emb_list}
\alias{apply_interleaved_rotary_emb_list}
\title{Apply interleaved rotary embedding}
\usage{
apply_interleaved_rotary_emb_list(x, freqs)
}
\description{
Apply interleaved rotary embedding
}
\keyword{internal}
