% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/unet_modules.R
\name{BasicTransformerBlock}
\alias{BasicTransformerBlock}
\title{Basic Transformer Block}
\usage{
BasicTransformerBlock(dim, n_heads, d_head, context_dim = NULL)
}
\description{
Basic Transformer Block
}
\keyword{internal}
