% tinyrox says don't edit this manually, but it can't stop you!
\name{CLIPTokenizer}
\alias{CLIPTokenizer}
\title{Tokenize a prompt}
\usage{
CLIPTokenizer(prompt,
              merges = system.file("tokenizer/merges.txt", package = "diffuseR"),
              vocab_file = system.file("tokenizer/vocab.json", package = "diffuseR"),
              pad_token = 0L)
}
\arguments{
\item{prompt}{A character string prompt describing the image to generate.}

\item{merges}{Path to the merges file (BPE merges).}

\item{vocab_file}{Path to the vocabulary file (token->id mapping).}

\item{pad_token}{The token ID used for padding (default is 0).}
}
\value{
A 2D torch tensor of shape c(1, 77) containing the token IDs.
}
\description{
Tokenize a prompt
}
