% tinyrox says don't edit this manually, but it can't stop you!
\name{create_sliding_window_mask}
\alias{create_sliding_window_mask}
\title{Create sliding window causal attention mask}
\usage{
create_sliding_window_mask(seq_len, window_size, device = "cpu")
}
\description{
Create sliding window causal attention mask
}
\keyword{internal}
