% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/unet.R
\name{detect_unet_architecture}
\alias{detect_unet_architecture}
\title{Detect UNet architecture from TorchScript file}
\usage{
detect_unet_architecture(torchscript_path)
}
\arguments{
\item{torchscript_path}{Path to TorchScript UNet .pt file}
}
\value{
List with architecture parameters
}
\description{
Detect UNet architecture from TorchScript file
}
\keyword{internal}
