% tinyrox says don't edit this manually, but it can't stop you!
\name{detect_unet_sdxl_architecture}
\alias{detect_unet_sdxl_architecture}
\title{Detect SDXL UNet architecture from TorchScript file}
\usage{
detect_unet_sdxl_architecture(torchscript_path)
}
\arguments{
\item{torchscript_path}{Path to TorchScript SDXL UNet .pt file}
}
\value{
List with architecture parameters
}
\description{
Detect SDXL UNet architecture from TorchScript file
}
\keyword{internal}
