% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionmap.R
\name{diffusionmap}
\alias{diffusionmap}
\title{Diffusion maps}
\usage{
diffusionmap(D, numeigen = 8, t = 0, maxdim = 50, epsilon = NULL,
  rsvd = TRUE)
}
\arguments{
\item{D}{distance matrix}

\item{numeigen}{number of diffusion coordinates}

\item{t}{length of markov chain run}

\item{maxdim}{default number of coordinates if numeigen NULL}

\item{epsilon}{value to use in kernel}

\item{rsvd}{true or false parameter to use rsvd or not}
}
\description{
Diffusion maps
}
