% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/save_video.R
\name{.ffmpeg_available}
\alias{.ffmpeg_available}
\title{Check if FFmpeg is Available}
\usage{
.ffmpeg_available()
}
\value{
Logical. TRUE if ffmpeg is in PATH.
}
\description{
Check if FFmpeg is Available
}
\keyword{internal}
