% tinyrox says don't edit this manually, but it can't stop you!
\name{.ltx23_release_attn_buffers}
\alias{.ltx23_release_attn_buffers}
\title{Release the attention scratch buffers}
\usage{
.ltx23_release_attn_buffers()
}
\value{
Invisibly, NULL.
}
\description{
Release the attention scratch buffers
}
\keyword{internal}
