% tinyrox says don't edit this manually, but it can't stop you!
\name{.verbosity}
\alias{.verbosity}
\title{Normalize a verbosity flag}
\usage{
.verbosity(verbose)
}
\arguments{
\item{verbose}{Logical, or one of "silent", "progress", "steps".
TRUE maps to "steps" and FALSE to "silent".}
}
\value{
One of "silent", "progress", "steps".
}
\description{
Normalize a verbosity flag
}
\keyword{internal}
