% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/vae_ltx2.R
\name{enable_tiling}
\alias{enable_tiling}
\title{Enable tiled encoding/decoding for memory efficiency}
\usage{
enable_tiling(
  tile_sample_min_height = NULL,
  tile_sample_min_width = NULL,
  tile_sample_min_num_frames = NULL,
  tile_sample_stride_height = NULL,
  tile_sample_stride_width = NULL,
  tile_sample_stride_num_frames = NULL
)
}
\description{
Enable tiled encoding/decoding for memory efficiency
}
