% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer_bpe.R
\name{encode_single}
\alias{encode_single}
\title{Encode a single text string}
\usage{
encode_single(tokenizer, text, add_special_tokens = TRUE)
}
\description{
Encode a single text string
}
\keyword{internal}
