% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2_modules.R
\name{feed_forward}
\alias{feed_forward}
\title{FeedForward module}
\usage{
feed_forward(
  dim,
  dim_out = NULL,
  mult = 4L,
  dropout = 0,
  activation_fn = "gelu-approximate",
  inner_dim = NULL,
  bias = TRUE
)
}
\description{
FeedForward module
}
\keyword{internal}
