% tinyrox says don't edit this manually, but it can't stop you!
\name{flux2_is_quant_key}
\alias{flux2_is_quant_key}
\title{Test whether a FLUX.2 key is in the quantization cast set}
\usage{
flux2_is_quant_key(key)
}
\arguments{
\item{key}{Character vector of parameter names (diffusers-style).}
}
\value{
Logical vector.
}
\description{
Test whether a FLUX.2 key is in the quantization cast set
}
