% tinyrox says don't edit this manually, but it can't stop you!
\name{gemma3_decoder_layer}
\alias{gemma3_decoder_layer}
\title{Gemma3 Decoder Layer}
\usage{
gemma3_decoder_layer(config, layer_idx = 0L)
}
\arguments{
\item{config}{Model configuration.}

\item{layer_idx}{Integer. Layer index.}
}
\description{
Single transformer block with pre-norm attention and MLP.
}
\keyword{internal}
