% tinyrox says don't edit this manually, but it can't stop you!
\name{gemma3_mlp}
\alias{gemma3_mlp}
\title{Gemma3 MLP}
\usage{
gemma3_mlp(config)
}
\arguments{
\item{config}{List with hidden_size and intermediate_size.}
}
\description{
Feed-forward network with gated linear units and GELU activation.
}
\keyword{internal}
