% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2devices.R
\name{get_required_components}
\alias{get_required_components}
\title{Get required components for each model type}
\usage{
get_required_components(model_name)
}
\description{
Get required components for each model type
}
