% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2_modules.R
\name{get_timestep_embedding}
\alias{get_timestep_embedding}
\title{Get timestep embedding (sinusoidal)}
\usage{
get_timestep_embedding(
  timesteps,
  embedding_dim,
  flip_sin_to_cos = FALSE,
  downscale_freq_shift = 1,
  scale = 1,
  max_period = 10000
)
}
\description{
Get timestep embedding (sinusoidal)
}
\keyword{internal}
