% tinyrox says don't edit this manually, but it can't stop you!
\name{greedy_tokenize}
\alias{greedy_tokenize}
\title{Greedy longest match tokenization}
\usage{
greedy_tokenize(text, vocab, byte_fallback = FALSE, unk_token = NULL)
}
\description{
Greedy longest match tokenization
}
\keyword{internal}
