% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelEpsilon.R
\name{kernelEpsilon}
\alias{kernelEpsilon}
\title{Function to compute the optimal epsilon}
\usage{
kernelEpsilon(D, p = 0.01)
}
\arguments{
\item{D}{distance matrix}

\item{p}{distance of neighbors}
}
\description{
Function to compute the optimal epsilon
}
