% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/gemma3_text_encoder.R
\name{load_gemma3_weights}
\alias{load_gemma3_weights}
\title{Load weights into Gemma3 model}
\usage{
load_gemma3_weights(model, weights, verbose = TRUE)
}
\description{
Load weights into Gemma3 model
}
\keyword{internal}
