% tinyrox says don't edit this manually, but it can't stop you!
\name{load_ltx2_connector_weights}
\alias{load_ltx2_connector_weights}
\title{Load weights into LTX2 connectors module}
\usage{
load_ltx2_connector_weights(connectors, weights, verbose = TRUE)
}
\arguments{
\item{connectors}{LTX2 connectors module}

\item{weights}{Named list of weight tensors}

\item{verbose}{Print progress}
}
\description{
Load weights into LTX2 connectors module
}
\keyword{internal}
