% tinyrox says don't edit this manually, but it can't stop you!
\name{ltx23_distilled_sigmas}
\alias{ltx23_distilled_sigmas}
\title{Official distilled sigma schedule}
\usage{
ltx23_distilled_sigmas()
}
\value{
Numeric vector of length 9.
}
\description{
The distilled LTX sigma values (with terminal zero appended), as
published in the Apache-2.0 diffusers reference
(pipelines/ltx2/utils.py).
}
