% tinyrox says don't edit this manually, but it can't stop you!
\name{ltx23_is_fp8_cast_key}
\alias{ltx23_is_fp8_cast_key}
\title{Test whether a mapped DiT key is in the official fp8 cast set}
\usage{
ltx23_is_fp8_cast_key(mapped_key)
}
\arguments{
\item{mapped_key}{Character vector of mapped (diffusers-style)
parameter names.}
}
\value{
Logical vector.
}
\description{
Test whether a mapped DiT key is in the official fp8 cast set
}
