% tinyrox says don't edit this manually, but it can't stop you!
\name{ltx23_map_connector_key}
\alias{ltx23_map_connector_key}
\title{Map an official connectors checkpoint key to the R module name}
\usage{
ltx23_map_connector_key(key)
}
\arguments{
\item{key}{Character. Checkpoint key.}
}
\value{
Character. Module parameter name.
}
\description{
Map an official connectors checkpoint key to the R module name
}
