% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2_modules.R
\name{ltx2_ada_layer_norm_single}
\alias{ltx2_ada_layer_norm_single}
\title{LTX2 AdaLayerNorm Single}
\usage{
ltx2_ada_layer_norm_single(
  embedding_dim,
  num_mod_params = 6L,
  use_additional_conditions = FALSE
)
}
\description{
LTX2 AdaLayerNorm Single
}
\keyword{internal}
