% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2.R
\name{ltx2_audio_video_rotary_pos_embed}
\alias{ltx2_audio_video_rotary_pos_embed}
\title{LTX2 Audio-Video Rotary Positional Embeddings}
\usage{
ltx2_audio_video_rotary_pos_embed(
  dim,
  patch_size = 1L,
  patch_size_t = 1L,
  base_num_frames = 20L,
  base_height = 2048L,
  base_width = 2048L,
  sampling_rate = 16000L,
  hop_length = 160L,
  scale_factors = c(8L, 32L, 32L),
  theta = 10000,
  causal_offset = 1L,
  modality = "video",
  double_precision = TRUE,
  rope_type = "interleaved",
  num_attention_heads = 32L
)
}
\description{
LTX2 Audio-Video Rotary Positional Embeddings
}
\keyword{internal}
