% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/text_encoder_ltx2.R
\name{ltx2_connector_transformer_1d}
\alias{ltx2_connector_transformer_1d}
\title{1D Connector Transformer for LTX2}
\usage{
ltx2_connector_transformer_1d(
  num_attention_heads = 30L,
  attention_head_dim = 128L,
  num_layers = 2L,
  num_learnable_registers = 128L,
  rope_base_seq_len = 4096L,
  rope_theta = 10000,
  rope_double_precision = TRUE,
  eps = 1e-06,
  causal_temporal_positioning = FALSE,
  rope_type = "interleaved"
)
}
\description{
1D Connector Transformer for LTX2
}
\keyword{internal}
