% tinyrox says don't edit this manually, but it can't stop you!
\name{ltx2_rotary_pos_embed_1d}
\alias{ltx2_rotary_pos_embed_1d}
\title{1D Rotary Position Embeddings for LTX2 Text Connectors}
\usage{
ltx2_rotary_pos_embed_1d(
  dim,
  base_seq_len = 4096L,
  theta = 10000,
  double_precision = TRUE,
  rope_type = "interleaved",
  num_attention_heads = 32L
)
}
\description{
1D Rotary Position Embeddings for LTX2 Text Connectors
}
\keyword{internal}
