% tinyrox says don't edit this manually, but it can't stop you!
\name{ltx2_transformer_block_1d}
\alias{ltx2_transformer_block_1d}
\title{1D Transformer Block for LTX2 Text Connectors}
\usage{
ltx2_transformer_block_1d(
  dim,
  num_attention_heads,
  attention_head_dim,
  activation_fn = "gelu-approximate",
  eps = 1e-06,
  rope_type = "interleaved"
)
}
\description{
1D Transformer Block for LTX2 Text Connectors
}
\keyword{internal}
