% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2_modules.R
\name{ltx2_video_transformer_block}
\alias{ltx2_video_transformer_block}
\title{LTX2 Video Transformer Block (Audio-Video)}
\usage{
ltx2_video_transformer_block(
  dim,
  num_attention_heads,
  attention_head_dim,
  cross_attention_dim,
  audio_dim,
  audio_num_attention_heads,
  audio_attention_head_dim,
  audio_cross_attention_dim,
  qk_norm = "rms_norm_across_heads",
  activation_fn = "gelu-approximate",
  attention_bias = TRUE,
  attention_out_bias = TRUE,
  eps = 1e-06,
  elementwise_affine = FALSE,
  rope_type = "interleaved"
)
}
\description{
LTX2 Video Transformer Block (Audio-Video)
}
\keyword{internal}
