% tinyrox says don't edit this manually, but it can't stop you!
\name{pixart_alpha_combined_timestep_size_embeddings}
\alias{pixart_alpha_combined_timestep_size_embeddings}
\title{PixArt Alpha Combined Timestep Size Embeddings}
\usage{
pixart_alpha_combined_timestep_size_embeddings(
  embedding_dim,
  size_emb_dim,
  use_additional_conditions = FALSE
)
}
\description{
PixArt Alpha Combined Timestep Size Embeddings
}
\keyword{internal}
