% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_image.R
\name{preprocess_image}
\alias{preprocess_image}
\title{Preprocess image for Stable Diffusion}
\usage{
preprocess_image(input, device = "cpu", width = 512, height = 512)
}
\arguments{
\item{input}{File path to .jpg or .png, or a 3D array}

\item{device}{Target device for torch ("cpu" or "cuda")}
}
\value{
Torch tensor of shape \link{1, 3, 512, 512}, scaled to \link{-1, 1}
}
\description{
Preprocess image for Stable Diffusion
}
