% tinyrox says don't edit this manually, but it can't stop you!
\name{print.bpe_tokenizer}
\alias{print.bpe_tokenizer}
\title{Print BPE Tokenizer}
\usage{
\method{print}{bpe_tokenizer}(x, ...)
}
\arguments{
\item{x}{A bpe_tokenizer object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}. Called for the side effect of
  printing a summary of the tokenizer to the console.
}
\description{
Print BPE Tokenizer
}
