% tinyrox says don't edit this manually, but it can't stop you!
\name{print.diffuseR_resident}
\alias{print.diffuseR_resident}
\title{Print a resident handle}
\usage{
\method{print}{diffuseR_resident}(x, ...)
}
\arguments{
\item{x}{A \code{diffuseR_resident} handle.}

\item{...}{Ignored.}
}
\value{
Invisibly \code{x}. Called for the side effect of printing a
  one-block summary to the console.
}
\description{
Print a resident handle
}
