% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/text_encoder.R
\name{quick_gelu}
\alias{quick_gelu}
\title{QuickGELU activation}
\usage{
quick_gelu(x)
}
\arguments{
\item{x}{Input tensor}
}
\description{
GELU approximation used by OpenAI CLIP: x * sigmoid(1.702 * x)
}
\keyword{internal}
