% tinyrox says don't edit this manually, but it can't stop you!
\name{qwen_bpe_tokenizer}
\alias{qwen_bpe_tokenizer}
\title{Load a Qwen2 byte-level BPE tokenizer}
\usage{
qwen_bpe_tokenizer(tokenizer_path)
}
\arguments{
\item{tokenizer_path}{Path to a tokenizer.json (or a directory
containing one).}
}
\value{
A \code{qwen_tokenizer} object.
}
\description{
Load a Qwen2 byte-level BPE tokenizer
}
