% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizedSVD.R
\name{randomizedSVD}
\alias{randomizedSVD}
\title{Randomized SVD}
\usage{
randomizedSVD(A, k = 6, p = 5, q = 2, sdist = "normal")
}
\arguments{
\item{A}{the matrix we want a low rank approx of}

\item{k}{dominant # eigenfunctions of an svd}

\item{p}{oversampling parameter}

\item{q}{number of subspace iterations}

\item{sampling_dist}{distribution to use with random matrix generation}
}
\description{
Randomized SVD
}
