% tinyrox says don't edit this manually, but it can't stop you!
\name{save_video_av}
\alias{save_video_av}
\title{Save Video using av Package}
\usage{
save_video_av(video, file, fps = 24, verbose = TRUE)
}
\arguments{
\item{video}{Array of video frames [T, H, W, C].}

\item{file}{Output file path.}

\item{fps}{Frames per second.}

\item{verbose}{Logical.}
}
\description{
Save Video using av Package
}
\keyword{internal}
