% tinyrox says don't edit this manually, but it can't stop you!
\name{SpatialTransformer}
\alias{SpatialTransformer}
\title{Spatial Transformer (Attention Block)}
\usage{
SpatialTransformer(
  in_channels,
  n_heads,
  d_head,
  depth = 1L,
  context_dim = NULL
)
}
\description{
Spatial Transformer (Attention Block)
}
\keyword{internal}
