% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2devices.R
\name{standardize_devices}
\alias{standardize_devices}
\title{Standardize devices configuration}
\usage{
standardize_devices(devices, required_components)
}
\description{
Standardize devices configuration
}
