% tinyrox says don't edit this manually, but it can't stop you!
\name{timesteps_module}
\alias{timesteps_module}
\title{Timesteps module}
\usage{
timesteps_module(
  num_channels,
  flip_sin_to_cos = TRUE,
  downscale_freq_shift = 0,
  scale = 1L
)
}
\description{
Timesteps module
}
\keyword{internal}
