% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/unet_modules.R
\name{UNetCrossAttention}
\alias{UNetCrossAttention}
\title{Cross-Attention for UNet}
\usage{
UNetCrossAttention(query_dim, context_dim = NULL, heads = 8L, dim_head = 64L)
}
\description{
Cross-Attention for UNet
}
\keyword{internal}
