% tinyrox says don't edit this manually, but it can't stop you!
\name{unigram_tokenizer}
\alias{unigram_tokenizer}
\title{Load a Unigram tokenizer from tokenizer.json}
\usage{
unigram_tokenizer(tokenizer_path)
}
\arguments{
\item{tokenizer_path}{Path to a HuggingFace tokenizer.json with a
Unigram model, or a directory containing one.}
}
\value{
A \code{unigram_tokenizer} object.
}
\description{
Load a Unigram tokenizer from tokenizer.json
}
