% tinyrox says don't edit this manually, but it can't stop you!
\name{upsampler_ltx2}
\alias{upsampler_ltx2}
\title{LTX-2 Latent Upsampler}
\description{
Spatial 2x upsampling of video latents using Conv3d ResBlocks and
PixelShuffle.  Used between Stage 1 (half-resolution) and Stage 2
(full-resolution) in the two-stage distilled pipeline.
}
