% tinyrox says don't edit this manually, but it can't stop you!
\name{VAEAttentionBlock}
\alias{VAEAttentionBlock}
\title{VAE Attention Block}
\usage{
VAEAttentionBlock(channels, norm_groups = 32)
}
\arguments{
\item{channels}{Number of channels}
}
\description{
Self-attention for VAE mid block
}
\keyword{internal}
