% tinyrox says don't edit this manually, but it can't stop you!
\name{VAEResnetBlock}
\alias{VAEResnetBlock}
\title{VAE ResNet Block}
\usage{
VAEResnetBlock(in_channels, out_channels, norm_groups = 32)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}
}
\description{
VAE ResNet Block
}
\keyword{internal}
