% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer_bpe.R
\name{vocab_size}
\alias{vocab_size}
\title{Get vocabulary size}
\usage{
vocab_size(tokenizer)
}
\arguments{
\item{tokenizer}{A bpe_tokenizer object.}
}
\value{
Integer vocabulary size.
}
\description{
Get vocabulary size
}
