% tinyrox says don't edit this manually, but it can't stop you!
\name{zimage_is_quant_key}
\alias{zimage_is_quant_key}
\title{Test whether a Z-Image key is in the quantization cast set}
\usage{
zimage_is_quant_key(key)
}
\arguments{
\item{key}{Character vector of parameter names (diffusers-style).}
}
\value{
Logical vector.
}
\description{
Test whether a Z-Image key is in the quantization cast set
}
