% tinyrox says don't edit this manually, but it can't stop you!
\name{zimage_pad_len}
\alias{zimage_pad_len}
\title{Padding length to the next multiple of 32}
\usage{
zimage_pad_len(n)
}
\arguments{
\item{n}{Integer token count.}
}
\value{
Integer pad length in [0, 31].
}
\description{
Padding length to the next multiple of 32
}
\keyword{internal}
