% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2_modules.R
\name{apply_interleaved_rotary_emb_list}
\alias{apply_interleaved_rotary_emb_list}
\title{Apply interleaved rotary embedding}
\usage{
apply_interleaved_rotary_emb_list(x, freqs)
}
\description{
Apply interleaved rotary embedding
}
\keyword{internal}
