% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/vae_ltx2.R
\name{diagonal_gaussian_distribution}
\alias{diagonal_gaussian_distribution}
\title{Diagonal Gaussian Distribution}
\usage{
diagonal_gaussian_distribution(parameters)
}
\arguments{
\item{parameters}{Tensor of concatenated mean and log variance.}
}
\description{
Represents a diagonal Gaussian distribution for VAE latents.
}
