% tinyrox says don't edit this manually, but it can't stop you!
\name{download_flux2}
\alias{download_flux2}
\title{Download and Prepare FLUX.2 Klein 4B Weights}
\description{
Downloads FLUX.2-klein-4B from HuggingFace (Apache-2.0, ungated) and
quantizes the 4B transformer to a local fp8 (~4 GB) or NF4 (~2.3 GB)
artifact.
}
