% tinyrox says don't edit this manually, but it can't stop you!
\name{gelu_activation}
\alias{gelu_activation}
\title{GELU activation with optional approximation}
\usage{
gelu_activation(dim_in, dim_out, approximate = "none", bias = TRUE)
}
\description{
GELU activation with optional approximation
}
\keyword{internal}
