% tinyrox says don't edit this manually, but it can't stop you!
\name{gemma3_config_ltx2}
\alias{gemma3_config_ltx2}
\title{Create Gemma3 configuration for LTX-2}
\usage{
gemma3_config_ltx2()
}
\value{
List with model configuration parameters.
}
\description{
Returns the default configuration used by LTX-2's text encoder.
}
