% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/models2devices.R
\name{get_required_components}
\alias{get_required_components}
\title{Get required components for each model type}
\usage{
get_required_components(model_name)
}
\arguments{
\item{model_name}{A character string representing the name of the model.}
}
\value{
A character vector of required components for the specified model.
}
\description{
This function returns a list of required components for each supported
model type.
}
