% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer_bpe.R
\name{greedy_tokenize}
\alias{greedy_tokenize}
\title{Greedy longest match tokenization}
\usage{
greedy_tokenize(text, vocab, byte_fallback = FALSE, unk_token = NULL)
}
\description{
Greedy longest match tokenization
}
\keyword{internal}
