% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2img.R
\name{img2img}
\alias{img2img}
\title{Image-to-Image Generation with Stable Diffusion}
\usage{
img2img(
  input_image,
  prompt,
  negative_prompt = NULL,
  img_dim = 512,
  model_name = "stable-diffusion-2-1",
  devices = "cpu",
  unet_dtype_str = "float16",
  scheduler = "ddim",
  num_inference_steps = 50,
  strength = 0.8,
  guidance_scale = 7.5,
  seed = NULL,
  save_to = NULL,
  metadata_path = NULL,
  ...
)
}
\arguments{
\item{input_image}{Path to the input image or a tensor representing the image.}

\item{prompt}{Text prompt to guide the image generation.}

\item{negative_prompt}{Optional negative prompt to guide the image generation.`}

\item{img_dim}{Dimension of the output image (default: 512).}

\item{model_name}{Name of the Stable Diffusion model to use (default: "stable-diffusion-2-1").}

\item{devices}{A named list of devices for each model component (e.g., \code{list(unet = "cuda", decoder = "cpu", text_encoder = "cpu", encoder = "cpu")}).}

\item{unet_dtype_str}{Optional A character for dtype of the unet component (typically "torch_float16" for cuda and "torch_float32" for cpu).}

\item{scheduler}{Scheduler to use for the diffusion process (default: "ddim").}

\item{num_inference_steps}{Number of diffusion steps (default: 50).}

\item{strength}{Strength of the image-to-image transformation (default: 0.8).}

\item{guidance_scale}{Scale for classifier-free guidance (default: 7.5).}

\item{seed}{Random seed for reproducibility (default: NULL).}

\item{save_to}{Path to save the generated image (default: NULL).}

\item{metadata_path}{Path to save metadata (default: NULL).}

\item{...}{Additional arguments for future use.}
}
\value{
Generated image tensor.
}
\description{
This function generates an image based on an input image and a text prompt using the Stable Diffusion model.
It allows for various configurations such as model name, device, scheduler, and more.
}
