% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/text_encoder_ltx2.R
\name{load_ltx2_connector_weights}
\alias{load_ltx2_connector_weights}
\title{Load weights into LTX2 connectors module}
\usage{
load_ltx2_connector_weights(connectors, weights, verbose = TRUE)
}
\arguments{
\item{connectors}{LTX2 connectors module}

\item{weights}{Named list of weight tensors}

\item{verbose}{Print progress}
}
\description{
Load weights into LTX2 connectors module
}
\keyword{internal}
