% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2.R
\name{load_ltx2_transformer_weights}
\alias{load_ltx2_transformer_weights}
\title{Load weights into LTX2 transformer module}
\usage{
load_ltx2_transformer_weights(transformer, weights, verbose = TRUE)
}
\arguments{
\item{transformer}{LTX2 transformer module}

\item{weights}{Named list of weight tensors}

\item{verbose}{Print progress}
}
\description{
Load weights into LTX2 transformer module
}
\keyword{internal}
