% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pipeline.R
\name{load_pipeline}
\alias{load_pipeline}
\title{Load a diffusion model pipeline}
\usage{
load_pipeline(model_name, m2d, i2i = FALSE, unet_dtype_str, ...)
}
\arguments{
\item{model_name}{The name of the model to load.}

\item{m2d}{A list containing model-to-device mappings and configurations.}

\item{i2i}{Logical indicating whether to load the encoder for img2img().}

\item{unet_dtype_str}{A string representing the data type for the UNet model (e.g., "float32", "float16").}

\item{...}{Additional arguments passed to the model loading functions.}
}
\value{
An environment containing the loaded models and configuration.
}
\description{
This function loads a diffusion model pipeline consisting of a UNet, VAE decoder, and text encoder.
It initializes the models and sets up the environment for inference.
}
\examples{
\dontrun{
pipeline <- load_pipeline("my_model", device = "cuda")
}

}
