% tinyrox says don't edit this manually, but it can't stop you!
\name{ltx23_map_audio_vae_key}
\alias{ltx23_map_audio_vae_key}
\title{Map an official audio VAE checkpoint key to the R module name}
\usage{
ltx23_map_audio_vae_key(key)
}
\arguments{
\item{key}{Character. Checkpoint key.}
}
\value{
Character.
}
\description{
Map an official audio VAE checkpoint key to the R module name
}
