% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models2devices.R
\name{models2devices}
\alias{models2devices}
\title{models2devices}
\usage{
models2devices(model_name, devices = "cpu", unet_dtype_str = NULL)
}
\arguments{
\item{model_name}{A character string representing the name of the model to be used.}

\item{devices}{A character string or a named list specifying the devices for different components of the model.}

\item{unet_dtype_str}{A character string specifying the data type for the UNet model.}
}
\value{
A list containing the model directory, model files, device configuration, UNet data type, and CPU/CUDA devices.
}
\description{
This function sets up the model directory, device configuration, and data types for the Stable Diffusion model.
It checks the validity of the model name and devices, and downloads the model if necessary.
It also sets the data type for the UNet model based on the device configuration.
}
