% tinyrox says don't edit this manually, but it can't stop you!
\name{pixart_alpha_text_projection}
\alias{pixart_alpha_text_projection}
\title{PixArt Alpha Text Projection}
\usage{
pixart_alpha_text_projection(
  in_features,
  hidden_size,
  out_features = NULL,
  act_fn = "gelu_tanh"
)
}
\description{
PixArt Alpha Text Projection
}
\keyword{internal}
