% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tokenizer_bpe.R
\name{print.bpe_tokenizer}
\alias{print.bpe_tokenizer}
\title{Print BPE Tokenizer}
\usage{
\method{print}{bpe_tokenizer}(x, ...)
}
\arguments{
\item{x}{A bpe_tokenizer object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print BPE Tokenizer
}
