% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2_modules.R
\name{timestep_embedding_module}
\alias{timestep_embedding_module}
\title{Timestep embedding MLP}
\usage{
timestep_embedding_module(
  in_channels,
  time_embed_dim,
  act_fn = "silu",
  out_dim = NULL
)
}
\description{
Timestep embedding MLP
}
\keyword{internal}
