% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/dit_ltx2_modules.R
\name{timesteps_module}
\alias{timesteps_module}
\title{Timesteps module}
\usage{
timesteps_module(
  num_channels,
  flip_sin_to_cos = TRUE,
  downscale_freq_shift = 0,
  scale = 1L
)
}
\description{
Timesteps module
}
\keyword{internal}
