% tinyrox says don't edit this manually, but it can't stop you!
\name{txt2img_flux}
\alias{txt2img_flux}
\title{FLUX.1 Text-to-Image Pipeline}
\description{
FLUX latent packing helpers and (in later phases) the schnell
text-to-image pipeline. Ported from the diffusers reference
implementation (Apache-2.0, src/diffusers/pipelines/flux/
pipeline_flux.py).
}
