% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/vae_decoder.R
\name{VAEAttentionBlock}
\alias{VAEAttentionBlock}
\title{VAE Attention Block}
\usage{
VAEAttentionBlock(channels)
}
\arguments{
\item{channels}{Number of channels}
}
\description{
Self-attention for VAE mid block
}
\keyword{internal}
