% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/vae_decoder.R
\name{VAEResnetBlock}
\alias{VAEResnetBlock}
\title{VAE ResNet Block}
\usage{
VAEResnetBlock(in_channels, out_channels)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}
}
\description{
VAE ResNet Block
}
\keyword{internal}
