% tinyrox says don't edit this manually, but it can't stop you!
\name{VAEUpBlock}
\alias{VAEUpBlock}
\title{VAE Up Block}
\usage{
VAEUpBlock(in_channels, out_channels, num_resnets = 3, add_upsample = TRUE,
           norm_groups = 32)
}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{num_resnets}{Number of resnet blocks (default 3)}

\item{add_upsample}{Whether to add upsampler}
}
\description{
VAE Up Block
}
\keyword{internal}
