% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/vae_decoder.R
\name{VAEUpBlock}
\alias{VAEUpBlock}
\title{VAE Up Block}
\arguments{
\item{in_channels}{Input channels}

\item{out_channels}{Output channels}

\item{num_resnets}{Number of resnet blocks (default 3)}

\item{add_upsample}{Whether to add upsampler}
}
\description{
VAE Up Block
}
\keyword{internal}
