\name{VAEUpBlock}
\alias{VAEUpBlock}
\title{VAE Up Block}
\description{VAE Up Block}
\arguments{
  \item{in_channels}{Input channels}
  \item{out_channels}{Output channels}
  \item{num_resnets}{Number of resnet blocks (default 3)}
  \item{add_upsample}{Whether to add upsampler}
}
\keyword{internal}
